extract ventilation text cards template
